extracted functions
